+++ example/android/app/src/main/kotlin/net/zonble/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package net.zonble.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}